+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.46.0] - 2024-10-29
 
 - Add `buf registry whoami` command, which checks if you are logged in to the Buf Schema
   Registry at a given domain.
@@ -1283,7 +1283,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.45.0...HEAD
+[v1.46.0]: https://github.com/bufbuild/buf/compare/v1.45.0...v1.46.0
 [v1.45.0]: https://github.com/bufbuild/buf/compare/v1.44.0...v1.45.0
 [v1.44.0]: https://github.com/bufbuild/buf/compare/v1.43.0...v1.44.0
 [v1.43.0]: https://github.com/bufbuild/buf/compare/v1.42.0...v1.43.0